Move StopWithDetails type into EnrichedStopRegistryTypes.ts
