+++ ui/src/components/stop-registry/stops/stop-details/basic-details/BasicDetailsSection.tsx
@@ -1,6 +1,7 @@
 import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import { StopWithDetails, useEditStopBasicDetails } from '../../../../../hooks';
+import { useEditStopBasicDetails } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 import { showSuccessToast, submitFormByRef } from '../../../../../utils';
 import { InfoContainer, useInfoContainerControls } from '../../../../common';
 import { stopInfoContainerColors } from '../stopInfoContainerColors';
