+++ ui/src/components/stop-registry/stops/stop-details/info-spots/SheltersInfoSpots.tsx
@@ -4,7 +4,7 @@ import {
   InfoSpotDetailsFragment,
   ShelterEquipmentDetailsFragment,
 } from '../../../../../generated/graphql';
-import { StopWithDetails } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 import { InfoSpotsSection } from './InfoSpotsSection';
 
 type Props = {
