+++ ui/src/components/stop-registry/stops/stop-details/maintenance/MaintenanceViewCard.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from 'react-i18next';
-import { StopWithDetails } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 import { MaintainerViewCard } from './MaintainerViewCard';
 import { getMaintainers } from './utils';
