+++ ui/src/components/stop-registry/stops/stop-details/shelters/SheltersSection.tsx
@@ -2,7 +2,8 @@ import compact from 'lodash/compact';
 import { useEffect, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ShelterEquipmentDetailsFragment } from '../../../../../generated/graphql';
-import { StopWithDetails, useEditStopShelters } from '../../../../../hooks';
+import { useEditStopShelters } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 import { showSuccessToast, submitFormByRef } from '../../../../../utils';
 import { InfoContainer, useInfoContainerControls } from '../../../../common';
 import { EmptyListHeaderButtons } from '../layout/EmptyListHeaderButtons';
