+++ ui/src/components/stop-registry/stops/stop-details/signage-details/SignageDetailsSection.tsx
@@ -1,9 +1,7 @@
 import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  StopWithDetails,
-  useEditStopSignageDetails,
-} from '../../../../../hooks';
+import { useEditStopSignageDetails } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 import { StopPlaceSignType } from '../../../../../types/stop-registry';
 import { showSuccessToast, submitFormByRef } from '../../../../../utils';
 import { InfoContainer, useInfoContainerControls } from '../../../../common';
