+++ ui/src/components/stop-registry/stops/stop-details/stop-version/utils/useCopyStopFormUtils.ts
@@ -3,8 +3,9 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useMemo } from 'react';
 import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import { StopWithDetails, useLoader } from '../../../../../../hooks';
+import { useLoader } from '../../../../../../hooks';
 import { Operation } from '../../../../../../redux';
+import { StopWithDetails } from '../../../../../../types';
 import { log, showToast } from '../../../../../../utils';
 import { getApolloErrorMessage } from '../../../../../../utils/apolloErrors';
 import {
