+++ ui/src/components/stop-registry/stops/stop-details/title-row/EditValidityButton.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { StopWithDetails } from '../../../../../hooks';
+import { StopWithDetails } from '../../../../../types';
 
 const DISABLE_UNTIL_IMPLEMENTED = true;
